Integration of Condorcet method poll (#79)
* replaced XPollService with a REST service

* Replaced XPollService with a REST service

* Replaced XPollService with a REST service

Used a script Service to expose additional functionalities of XPollManager. Changed REST URL. Changed the request parameters sent and expected. Modified the UI in accordance to the aformentioned changes.

* Replaced XPollService with a REST service

Code improvements

* Edited parent pom to have the API as module

* Replaced XPollService with REST service

Added Unit Tests for DefaultXPollManager. Code improvements.

* Changed XPollRecource javadoc

* Code improvements

* Code improvements

* Code improvements

* Cleaned imports in XPollScriptService

* Add support for a Condorcet method poll #76

Added the UI elements to enable the user to vote in an condorcet manner. The votes are currently handled as the traditional multiple choice way.

* Line separators

* Add support for a Condorcet method poll #76

Added backend implementation for determining the winner in a Condorcet Manner. Changed the event listener. Added one more unit test. Exposed one more function through the XPoll Service.

* Add support for a Condorcet method poll #76

Modified the EventListener once again. It didn't update the winner when the document was modified because, I believe, the wrong context was used to change the object value.

* Add support for a Condorcet method poll #76

Fixed a bug in which the Type field of the XPollClass was displayed and selected as null - caused by the change of the field from boolean to static list. The solution was to manually change the Template for XPoll Pages.

* Removed template name

* Added javadocs

* Add support for a Condorcet method poll #76

Set the PollResultsCalculator as a component with 4 implementations. One for each type of poll - two of them extending a fourth, Default, implementation. Created an abstract blueprint for all the calculators. Several code improvements.

* Add support for a Condorcet method poll #76

Code improvements.

* Add support for a Condorcet method poll #76

Implemented a way to migrate the old polls after updating the application. Code improvements,

* Add support for a Condorcet method poll #76

Code improvement.

* Add support for a Condorcet method poll #76

Code improvements

* Add support for a Condorcet method poll #76

Code improvements