Add support for a Condorcet method poll #76
Fixed a bug in which refreshing the page wouldn't display the previously selected options of the user (condorcet poll).
Fixed a bug in which the user needed to save the poll twice for the winner of the poll to be updated.
Modified the translation key of the xpollVotes because the condorcet poll calculates a score for each proposal rather than simply count the votes.
Fixed a bug in which having proposals with trailing spaces in a condorcet poll would cause NullPointerException.